requirements.txt
asgiref==3.7.2
attrs==23.1.0
dj-config-url==0.1.1
Django==4.2.2
django-filter==23.2
django-pyproject==1.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jsonschema==4.17.3
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyproject-toml==0.0.11
pyrsistent==0.19.3
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
tzdata==2023.3
xlwt==1.3.0
